requirements.txt
Django==4.2
django-environ==0.10.0
djangorestframework==3.14.0
django-silk==5.0.3
psycopg==3.1.9
celery==5.3.0
redis==4.5.5
pdfkit==1.0.0
requests==2.31.0
PyPDF2==3.0.1
boto3==1.26.161
django-storages==1.13.2
pytest-django==4.5.2
pytest==7.4.0
moto==4.2.0
django-dynamic-preferences==1.15.0
